Refactoring code

+++ cart/services.py
@@ -1,7 +1,4 @@
-from decimal import Decimal
-
-
-def replace_to_dot(price: float) -> Decimal:
+def replace_to_dot(price: float) -> float:
     for i in price.replace(',', '.').split():
         price = i
     return price
